ai-agent-hub/requirements.txt
fastapi
uvicorn[standard]
pyyaml
httpx
fastapi>=0.110
uvicorn[standard]>=0.30
pandas>=2.2
numpy>=1.26
scikit-learn>=1.4
shap>=0.45
fpdf2>=2.7
python-multipart>=0.0.9
pydantic>=2.6
aiofiles>=23.2
requests>=2.32
streamlit>=1.35
altair>=5.3
plotly>=5.18.0
statsmodels>=0.14
Pillow>=10.3
